MCTS_WebAgent-master/backend/requirements.txt
# backend/requirements.txt

fastapi
uvicorn
langchain
langchain-ollama
pydantic
requests
beautifulsoup4
lxml
numpy
python-dotenv
